feat: add brij logging
